chore(router): delete unused files and code & reorganize test components
