+++ packages/router/test/e2e/doc-example/src/app.ts
@@ -2,8 +2,8 @@ import { inject } from '@aurelia/kernel';
 import { customElement } from '@aurelia/runtime';
 import { Router } from '../../../../../router/src/index';
 import { About } from './components/about';
-import { Authors } from './components/authors';
-import { Books } from './components/books';
+import { Authors } from './components/authors/authors';
+import { Books } from './components/books/books';
 
 @inject(Router)
 @customElement({
